Move from `appdirs` to `platformdirs`
