+++ pyproject.toml
@@ -41,7 +41,6 @@ license-files = [
     "LICENSE",
 ]
 dependencies = [
-    "appdirs>=1.4.4",
     "app-model>=0.5.0,<0.6.0",
     "cachey>=0.2.1",
     "certifi>=2018.1.18",
@@ -58,6 +57,7 @@ dependencies = [
     "pandas>=1.3.3",
     "Pillow>=10.0",
     "pint>=0.17",
+    "platformdirs>=3.8.0",
     "psutil>=5.9.0",
     "psygnal>=0.14.2",
     "pydantic>=2.8.0",
