+++ src/napari/resources/_icons.py
@@ -4,7 +4,7 @@ from functools import lru_cache
 from itertools import product
 from pathlib import Path
 
-from napari.utils._appdirs import user_cache_dir
+from napari.utils._platformdirs import user_cache_dir
 from napari.utils.translations import trans
 
 LOADING_GIF_PATH = str((Path(__file__).parent / 'loading.gif').resolve())
